chore: replace `useModalModelVersionEdit` with `ModelVersionEditModal`
